break apart staff api
