+++ seqr/views/apis/report_api.py
@@ -1,135 +1,31 @@
-import base64
 from collections import defaultdict
-import json
 import logging
-import re
 import requests
-import urllib3
 
 from datetime import datetime, timedelta
 from dateutil import relativedelta as rdelta
 from django.contrib.admin.views.decorators import staff_member_required
-from django.contrib.postgres.aggregates import ArrayAgg
-from django.db.models import prefetch_related_objects, Q, Prefetch, Max
-from django.http.response import HttpResponse
+from django.db.models import Prefetch
 from django.utils import timezone
-from django.views.decorators.csrf import csrf_exempt
-from requests.exceptions import ConnectionError as RequestConnectionError
 
-from seqr.utils.elasticsearch.utils import get_es_client, get_index_metadata
-from seqr.utils.file_utils import file_iter
 from seqr.utils.gene_utils import get_genes
 from seqr.utils.xpos_utils import get_chrom_pos
 
-from matchmaker.matchmaker_utils import get_mme_genes_phenotypes_for_submissions, parse_mme_features, \
-    parse_mme_gene_variants, get_mme_metrics
-from seqr.views.apis.saved_variant_api import _add_locus_lists
 from seqr.views.utils.export_utils import export_multiple_files
-from seqr.views.utils.file_utils import parse_file
-from seqr.views.utils.json_utils import create_json_response, _to_camel_case
-from seqr.views.utils.orm_to_json_utils import _get_json_for_individuals, get_json_for_saved_variants, \
-    get_json_for_variant_functional_data_tag_types, get_json_for_projects, _get_json_for_families, \
-    get_json_for_locus_lists, _get_json_for_models, get_json_for_matchmaker_submissions, \
-    get_json_for_saved_variants_with_tags
-from seqr.views.utils.variant_utils import saved_variant_genes
+from seqr.views.utils.json_utils import create_json_response
+from seqr.views.utils.orm_to_json_utils import get_json_for_saved_variants
 
 from matchmaker.models import MatchmakerSubmission
-from seqr.models import Project, Family, VariantTag, VariantTagType, Sample, SavedVariant, Individual, ProjectCategory, \
-    LocusList
+from seqr.models import Project, Family, VariantTag, VariantTagType, Sample, SavedVariant, Individual, ProjectCategory
 from reference_data.models import Omim, HumanPhenotypeOntology
 
-from settings import ELASTICSEARCH_SERVER, KIBANA_SERVER, API_LOGIN_REQUIRED_URL, AIRTABLE_API_KEY, AIRTABLE_URL, \
-    KIBANA_ELASTICSEARCH_PASSWORD
+from settings import API_LOGIN_REQUIRED_URL, AIRTABLE_API_KEY, AIRTABLE_URL
 
 logger = logging.getLogger(__name__)
 
 HET = 'Heterozygous'
 HOM_ALT = 'Homozygous'
 HEMI = 'Hemizygous'
-
-MAX_SAVED_VARIANTS = 10000
-
-
-@staff_member_required(login_url=API_LOGIN_REQUIRED_URL)
-def elasticsearch_status(request):
-    client = get_es_client()
-
-    disk_fields = ['node', 'shards', 'disk.avail', 'disk.used', 'disk.percent']
-    disk_status = [{
-        _to_camel_case(field.replace('.', '_')): disk[field] for field in disk_fields
-    } for disk in client.cat.allocation(format="json", h=','.join(disk_fields))]
-
-    index_fields = ['index', 'docs.count', 'store.size', 'creation.date.string']
-    indices = [{
-        _to_camel_case(field.replace('.', '_')): index[field] for field in index_fields
-    } for index in client.cat.indices(format="json", h=','.join(index_fields))
-        if all(not index['index'].startswith(omit_prefix) for omit_prefix in ['.', 'index_operations_log'])]
-
-    aliases = defaultdict(list)
-    for alias in client.cat.aliases(format="json", h='alias,index'):
-        aliases[alias['alias']].append(alias['index'])
-
-    index_metadata = get_index_metadata('_all', client, use_cache=False)
-
-    active_samples = Sample.objects.filter(is_active=True).select_related('individual__family__project')
-
-    seqr_index_projects = defaultdict(lambda: defaultdict(set))
-    es_projects = set()
-    for sample in active_samples:
-        for index_name in sample.elasticsearch_index.split(','):
-            project = sample.individual.family.project
-            es_projects.add(project)
-            if index_name in aliases:
-                for aliased_index_name in aliases[index_name]:
-                    seqr_index_projects[aliased_index_name][project].add(sample.individual.guid)
-            else:
-                seqr_index_projects[index_name.rstrip('*')][project].add(sample.individual.guid)
-
-    for index in indices:
-        index_name = index['index']
-        index.update(index_metadata[index_name])
-
-        projects_for_index = []
-        for index_prefix in list(seqr_index_projects.keys()):
-            if index_name.startswith(index_prefix):
-                projects_for_index += list(seqr_index_projects.pop(index_prefix).keys())
-        index['projects'] = [{'projectGuid': project.guid, 'projectName': project.name} for project in projects_for_index]
-
-    errors = ['{} does not exist and is used by project(s) {}'.format(
-        index, ', '.join(['{} ({} samples)'.format(p.name, len(indivs)) for p, indivs in project_individuals.items()])
-    ) for index, project_individuals in seqr_index_projects.items() if project_individuals]
-
-    return create_json_response({
-        'indices': indices,
-        'diskStats': disk_status,
-        'elasticsearchHost': ELASTICSEARCH_SERVER,
-        'errors': errors,
-    })
-
-
-@staff_member_required(login_url=API_LOGIN_REQUIRED_URL)
-def mme_details(request):
-    submissions = MatchmakerSubmission.objects.filter(deleted_date__isnull=True)
-
-    hpo_terms_by_id, genes_by_id, gene_symbols_to_ids = get_mme_genes_phenotypes_for_submissions(submissions)
-
-    submission_json = get_json_for_matchmaker_submissions(
-        submissions, additional_model_fields=['label'], all_parent_guids=True)
-    submissions_by_guid = {s['submissionGuid']: s for s in submission_json}
-
-    for submission in submissions:
-        gene_variants = parse_mme_gene_variants(submission.genomic_features, gene_symbols_to_ids)
-        submissions_by_guid[submission.guid].update({
-            'phenotypes': parse_mme_features(submission.features, hpo_terms_by_id),
-            'geneVariants': gene_variants,
-            'geneSymbols': ','.join({genes_by_id.get(gv['geneId'], {}).get('geneSymbol') for gv in gene_variants})
-        })
-
-    return create_json_response({
-        'metrics': get_mme_metrics(),
-        'submissions': list(submissions_by_guid.values()),
-        'genesById': genes_by_id,
-    })
 
 
 @staff_member_required(login_url=API_LOGIN_REQUIRED_URL)
@@ -813,28 +709,6 @@ def discovery_sheet(request, project_guid):
     })
 
 
-@staff_member_required(login_url=API_LOGIN_REQUIRED_URL)
-def success_story(request, success_story_types):
-    if success_story_types == 'all':
-        families = Family.objects.filter(success_story__isnull=False)
-    else:
-        success_story_types = success_story_types.split(',')
-        families = Family.objects.filter(success_story_types__overlap=success_story_types)
-
-    rows = [{
-        "project_guid": family.project.guid,
-        "family_guid": family.guid,
-        "family_id": family.family_id,
-        "success_story_types": family.success_story_types,
-        "success_story": family.success_story,
-        "row_id": family.guid,
-    } for family in families]
-
-    return create_json_response({
-        'rows': rows,
-    })
-
-
 def _get_loaded_samples_by_family(project):
     loaded_samples = Sample.objects.filter(individual__family__project=project).select_related(
         'individual__family').order_by('loaded_date')
@@ -1227,332 +1101,3 @@ def _update_initial_omim_numbers(rows):
     for row in rows:
         if omim_number_map.get(row['omim_number_initial']):
             row['omim_number_initial'] = omim_number_map[row['omim_number_initial']]
-
-
-@staff_member_required(login_url=API_LOGIN_REQUIRED_URL)
-def saved_variants_page(request, tag):
-    gene = request.GET.get('gene')
-
-    if tag == 'ALL':
-        saved_variant_models = SavedVariant.objects.exclude(varianttag=None)
-    else:
-        tag_type = VariantTagType.objects.get(name=tag, project__isnull=True)
-        saved_variant_models = SavedVariant.objects.filter(varianttag__variant_tag_type=tag_type)
-
-    if gene:
-        saved_variant_models = saved_variant_models.filter(saved_variant_json__transcripts__has_key=gene)
-    elif saved_variant_models.count() > MAX_SAVED_VARIANTS:
-        return create_json_response({'error': 'Select a gene to filter variants'}, status=400)
-
-    prefetch_related_objects(saved_variant_models, 'family__project')
-    response_json = get_json_for_saved_variants_with_tags(saved_variant_models, add_details=True, include_missing_variants=True)
-
-    project_models_by_guid = {variant.family.project.guid: variant.family.project for variant in saved_variant_models}
-    families = {variant.family for variant in saved_variant_models}
-    individuals = Individual.objects.filter(family__in=families)
-
-    saved_variants = list(response_json['savedVariantsByGuid'].values())
-    genes = saved_variant_genes(saved_variants)
-    locus_lists_by_guid = _add_locus_lists(list(project_models_by_guid.values()), genes, include_all_lists=True)
-
-    projects_json = get_json_for_projects(list(project_models_by_guid.values()), user=request.user, add_project_category_guids_field=False)
-    functional_tag_types = get_json_for_variant_functional_data_tag_types()
-
-    variant_tag_types = VariantTagType.objects.filter(Q(project__in=project_models_by_guid.values()) | Q(project__isnull=True))
-    prefetch_related_objects(variant_tag_types, 'project')
-    variant_tags_json = _get_json_for_models(variant_tag_types)
-    tag_projects = {vt.guid: vt.project.guid for vt in variant_tag_types if vt.project}
-
-    for project_json in projects_json:
-        project_guid = project_json['projectGuid']
-        project_variant_tags = [
-            vt for vt in variant_tags_json if tag_projects.get(vt['variantTagTypeGuid'], project_guid) == project_guid]
-        project_json.update({
-            'locusListGuids': list(locus_lists_by_guid.keys()),
-            'variantTagTypes': sorted(project_variant_tags, key=lambda variant_tag_type: variant_tag_type['order'] or 0),
-            'variantFunctionalTagTypes': functional_tag_types,
-        })
-
-    families_json = _get_json_for_families(list(families), user=request.user, add_individual_guids_field=True)
-    individuals_json = _get_json_for_individuals(individuals, add_hpo_details=True, user=request.user)
-    for locus_list in get_json_for_locus_lists(LocusList.objects.filter(guid__in=locus_lists_by_guid.keys()), request.user):
-        locus_lists_by_guid[locus_list['locusListGuid']].update(locus_list)
-
-    response_json.update({
-        'genesById': genes,
-        'projectsByGuid': {project['projectGuid']: project for project in projects_json},
-        'familiesByGuid': {family['familyGuid']: family for family in families_json},
-        'individualsByGuid': {indiv['individualGuid']: indiv for indiv in individuals_json},
-        'locusListsByGuid': locus_lists_by_guid,
-    })
-    return create_json_response(response_json)
-
-
-@staff_member_required(login_url=API_LOGIN_REQUIRED_URL)
-def upload_qc_pipeline_output(request):
-    file_path = json.loads(request.body)['file']
-    raw_records = parse_file(file_path, file_iter(file_path))
-
-    json_records = [dict(zip(raw_records[0], row)) for row in raw_records[1:]]
-
-    try:
-        dataset_type, data_type, records_by_sample_id = _parse_raw_qc_records(json_records)
-    except ValueError as e:
-        return create_json_response({'errors': [str(e)]}, status=400, reason=str(e))
-
-    info_message = 'Parsed {} {} samples'.format(
-        len(json_records), 'SV' if dataset_type == Sample.DATASET_TYPE_SV_CALLS else data_type)
-    logger.info(info_message)
-    info = [info_message]
-    warnings = []
-
-    samples = Sample.objects.filter(
-        sample_id__in=records_by_sample_id.keys(),
-        sample_type=Sample.SAMPLE_TYPE_WES if data_type == 'exome' else Sample.SAMPLE_TYPE_WGS,
-        dataset_type=dataset_type,
-    ).exclude(
-        individual__family__project__name__in=EXCLUDE_PROJECTS
-    ).exclude(individual__family__project__projectcategory__name=EXCLUDE_PROJECT_CATEGORY)
-
-    sample_individuals = {
-        agg['sample_id']: agg['individuals'] for agg in
-        samples.values('sample_id').annotate(individuals=ArrayAgg('individual_id', distinct=True))
-    }
-
-    sample_individual_max_loaded_date = {
-        agg['individual_id']: agg['max_loaded_date'] for agg in
-        samples.values('individual_id').annotate(max_loaded_date=Max('loaded_date'))
-    }
-    individual_latest_sample_id = {
-        s.individual_id: s.sample_id for s in samples
-        if s.loaded_date == sample_individual_max_loaded_date.get(s.individual_id)
-    }
-
-    for sample_id, record in records_by_sample_id.items():
-        record['individual_ids'] = list({
-            individual_id for individual_id in sample_individuals.get(sample_id, [])
-            if individual_latest_sample_id[individual_id] == sample_id
-        })
-
-    missing_sample_ids = {sample_id for sample_id, record in records_by_sample_id.items() if not record['individual_ids']}
-    if missing_sample_ids:
-        individuals = Individual.objects.filter(individual_id__in=missing_sample_ids).exclude(
-            family__project__name__in=EXCLUDE_PROJECTS).exclude(
-            family__project__projectcategory__name=EXCLUDE_PROJECT_CATEGORY).filter(
-            sample__sample_type=Sample.SAMPLE_TYPE_WES if data_type == 'exome' else Sample.SAMPLE_TYPE_WGS).distinct()
-        individual_db_ids_by_id = defaultdict(list)
-        for individual in individuals:
-            individual_db_ids_by_id[individual.individual_id].append(individual.id)
-        for sample_id, record in records_by_sample_id.items():
-            if not record['individual_ids'] and len(individual_db_ids_by_id[sample_id]) >= 1:
-                record['individual_ids'] = individual_db_ids_by_id[sample_id]
-                missing_sample_ids.remove(sample_id)
-
-    multi_individual_samples = {
-        sample_id: len(record['individual_ids']) for sample_id, record in records_by_sample_id.items()
-        if len(record['individual_ids']) > 1}
-    if multi_individual_samples:
-        logger.info('Found {} multi-individual samples from qc output'.format(len(multi_individual_samples)))
-        warnings.append('The following {} samples were added to multiple individuals: {}'.format(
-            len(multi_individual_samples), ', '.join(
-                sorted(['{} ({})'.format(sample_id, count) for sample_id, count in multi_individual_samples.items()]))))
-
-    if missing_sample_ids:
-        logger.info('Missing {} samples from qc output'.format(len(missing_sample_ids)))
-        warnings.append('The following {} samples were skipped: {}'.format(
-            len(missing_sample_ids), ', '.join(sorted(list(missing_sample_ids)))))
-
-    records_with_individuals = [
-        record for sample_id, record in records_by_sample_id.items() if sample_id not in missing_sample_ids
-    ]
-
-    if dataset_type == Sample.DATASET_TYPE_SV_CALLS:
-        _update_individuals_sv_qc(records_with_individuals, request.user)
-    else:
-        _update_individuals_variant_qc(records_with_individuals, data_type, warnings, request.user)
-
-    message = 'Found and updated matching seqr individuals for {} samples'.format(len(json_records) - len(missing_sample_ids))
-    info.append(message)
-    logger.info(message)
-
-    return create_json_response({
-        'errors': [],
-        'warnings': warnings,
-        'info': info,
-    })
-
-
-def _parse_raw_qc_records(json_records):
-    # Parse SV QC
-    if all(field in json_records[0] for field in ['sample', 'lt100_raw_calls', 'lt10_highQS_rare_calls']):
-        records_by_sample_id = {
-            re.search('(\d+)_(?P<sample_id>.+)_v\d_Exome_GCP', record['sample']).group('sample_id'): record
-            for record in json_records}
-        return Sample.DATASET_TYPE_SV_CALLS, 'exome', records_by_sample_id
-
-    # Parse regular variant QC
-    missing_columns = [field for field in ['seqr_id', 'data_type', 'filter_flags', 'qc_metrics_filters', 'qc_pop']
-                       if field not in json_records[0]]
-    if missing_columns:
-        raise ValueError('The following required columns are missing: {}'.format(', '.join(missing_columns)))
-
-    data_types = {record['data_type'].lower() for record in json_records if record['data_type'].lower() != 'n/a'}
-    if len(data_types) == 0:
-        raise ValueError('No data type detected')
-    elif len(data_types) > 1:
-        raise ValueError('Multiple data types detected: {}'.format(' ,'.join(sorted(data_types))))
-    elif list(data_types)[0] not in DATA_TYPE_MAP:
-        message = 'Unexpected data type detected: "{}" (should be "exome" or "genome")'.format(list(data_types)[0])
-        raise ValueError(message)
-
-    data_type = DATA_TYPE_MAP[list(data_types)[0]]
-    records_by_sample_id = {record['seqr_id']: record for record in json_records}
-
-    return Sample.DATASET_TYPE_VARIANT_CALLS, data_type, records_by_sample_id
-
-
-def _update_individuals_variant_qc(json_records, data_type, warnings, user):
-    unknown_filter_flags = set()
-    unknown_pop_filter_flags = set()
-
-    inidividuals_by_population = defaultdict(list)
-    for record in json_records:
-        filter_flags = {}
-        for flag in json.loads(record['filter_flags']):
-            flag = '{}_{}'.format(flag, data_type) if flag == 'coverage' else flag
-            flag_col = FILTER_FLAG_COL_MAP.get(flag, flag)
-            if flag_col in record:
-                filter_flags[flag] = record[flag_col]
-            else:
-                unknown_filter_flags.add(flag)
-
-        pop_platform_filters = {}
-        for flag in json.loads(record['qc_metrics_filters']):
-            flag_col = 'sample_qc.{}'.format(flag)
-            if flag_col in record:
-                pop_platform_filters[flag] = record[flag_col]
-            else:
-                unknown_pop_filter_flags.add(flag)
-
-        if filter_flags or pop_platform_filters:
-            Individual.bulk_update(user, {
-                'filter_flags': filter_flags or None, 'pop_platform_filters': pop_platform_filters or None,
-            }, id__in=record['individual_ids'])
-
-        inidividuals_by_population[record['qc_pop'].upper()] += record['individual_ids']
-
-    for population, indiv_ids in inidividuals_by_population.items():
-        Individual.bulk_update(user, {'population': population}, id__in=indiv_ids)
-
-    if unknown_filter_flags:
-        message = 'The following filter flags have no known corresponding value and were not saved: {}'.format(
-            ', '.join(unknown_filter_flags))
-        logger.info(message)
-        warnings.append(message)
-
-    if unknown_pop_filter_flags:
-        message = 'The following population platform filters have no known corresponding value and were not saved: {}'.format(
-            ', '.join(unknown_pop_filter_flags))
-        logger.info(message)
-        warnings.append(message)
-
-
-def _update_individuals_sv_qc(json_records, user):
-    inidividuals_by_qc = defaultdict(list)
-    for record in json_records:
-        inidividuals_by_qc[(record['lt100_raw_calls'], record['lt10_highQS_rare_calls'])] += record['individual_ids']
-
-    for raw_flags, indiv_ids in inidividuals_by_qc.items():
-        lt100_raw_calls, lt10_highQS_rare_calls = raw_flags
-        sv_flags = []
-        if lt100_raw_calls == 'FALSE':
-            sv_flags.append('raw_calls:_>100')
-        if lt10_highQS_rare_calls == 'FALSE':
-            sv_flags.append('high_QS_rare_calls:_>10')
-        Individual.bulk_update(user, {'sv_flags': sv_flags or None}, id__in=indiv_ids)
-
-
-FILTER_FLAG_COL_MAP = {
-    'callrate': 'filtered_callrate',
-    'contamination': 'PCT_CONTAMINATION',
-    'chimera': 'AL_PCT_CHIMERAS',
-    'coverage_exome': 'HS_PCT_TARGET_BASES_20X',
-    'coverage_genome': 'WGS_MEAN_COVERAGE'
-}
-
-DATA_TYPE_MAP = {
-    'exome': 'exome',
-    'genome': 'genome',
-    'wes': 'exome',
-    'wgs': 'genome',
-}
-
-EXCLUDE_PROJECTS = [
-    '[DISABLED_OLD_CMG_Walsh_WES]', 'Old Engle Lab All Samples 352S', 'Old MEEI Engle Samples',
-    'kl_temp_manton_orphan-diseases_cmg-samples_exomes_v1', 'Interview Exomes', 'v02_loading_test_project',
-]
-EXCLUDE_PROJECT_CATEGORY = 'Demo'
-
-# Hop-by-hop HTTP response headers shouldn't be forwarded.
-# More info at: http://www.w3.org/Protocols/rfc2616/rfc2616-sec13.html#sec13.5.1
-EXCLUDE_HTTP_RESPONSE_HEADERS = {
-    'connection', 'keep-alive', 'proxy-authenticate',
-    'proxy-authorization', 'te', 'trailers', 'transfer-encoding', 'upgrade',
-}
-
-urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-
-
-@staff_member_required(login_url=API_LOGIN_REQUIRED_URL)
-@csrf_exempt
-def proxy_to_kibana(request):
-    headers = _convert_django_meta_to_http_headers(request.META)
-    headers['Host'] = KIBANA_SERVER
-    if KIBANA_ELASTICSEARCH_PASSWORD:
-        token = base64.b64encode('kibana:{}'.format(KIBANA_ELASTICSEARCH_PASSWORD).encode('utf-8'))
-        headers['Authorization'] = 'Basic {}'.format(token.decode('utf-8'))
-
-    url = "{scheme}://{host}{path}".format(scheme=request.scheme, host=KIBANA_SERVER, path=request.get_full_path())
-
-    request_method = getattr(requests.Session(), request.method.lower())
-
-    try:
-        # use stream=True because kibana returns gziped responses, and this prevents the requests module from
-        # automatically unziping them
-        response = request_method(url, headers=headers, data=request.body, stream=True, verify=True)
-        response_content = response.raw.read()
-        # make sure the connection is released back to the connection pool
-        # (based on http://docs.python-requests.org/en/master/user/advanced/#body-content-workflow)
-        response.close()
-
-        proxy_response = HttpResponse(
-            content=response_content,
-            status=response.status_code,
-            reason=response.reason,
-            charset=response.encoding
-        )
-
-        for key, value in response.headers.items():
-            if key.lower() not in EXCLUDE_HTTP_RESPONSE_HEADERS:
-                proxy_response[key.title()] = value
-
-        return proxy_response
-    except (ConnectionError, RequestConnectionError) as e:
-        logger.error(str(e))
-        return HttpResponse("Error: Unable to connect to Kibana {}".format(e), status=400)
-
-
-def _convert_django_meta_to_http_headers(request_meta_dict):
-    """Converts django request.META dictionary into a dictionary of HTTP headers."""
-
-    def convert_key(key):
-        # converting Django's all-caps keys (eg. 'HTTP_RANGE') to regular HTTP header keys (eg. 'Range')
-        return key.replace("HTTP_", "").replace('_', '-').title()
-
-    http_headers = {
-        convert_key(key): str(value).lstrip()
-        for key, value in request_meta_dict.items()
-        if key.startswith("HTTP_") or (key in ('CONTENT_LENGTH', 'CONTENT_TYPE') and value)
-    }
-
-    return http_headers
