+++ seqr/views/apis/summary_data_api_tests.py
@@ -1,0 +1,86 @@
+from datetime import datetime
+from django.urls.base import reverse
+import mock
+
+from seqr.views.apis.summary_data_api import mme_details, success_story, saved_variants_page
+from seqr.views.utils.test_utils import AuthenticationTestCase
+
+
+PROJECT_GUID = 'R0001_1kg'
+
+EXPECTED_SUCCESS_STORY = {'project_guid': 'R0001_1kg', 'family_guid': 'F000013_13', 'success_story_types': ['A'], 'family_id': 'no_individuals', 'success_story': 'Treatment is now available on compassionate use protocol (nucleoside replacement protocol)', 'row_id': 'F000013_13'}
+
+EXPECTED_MME_DETAILS_METRICS = {
+    u'numberOfPotentialMatchesSent': 1,
+    u'numberOfUniqueGenes': 4,
+    u'numberOfCases': 3,
+    u'numberOfRequestsReceived': 3,
+    u'numberOfSubmitters': 2,
+    u'numberOfUniqueFeatures': 4,
+    u'dateGenerated': '2020-04-27'
+}
+
+class SummaryDataAPITest(AuthenticationTestCase):
+    fixtures = ['users', '1kg_project', 'reference_data']
+
+    @mock.patch('matchmaker.matchmaker_utils.datetime')
+    def test_mme_details(self, mock_datetime):
+        url = reverse(mme_details)
+        self.check_staff_login(url)
+
+        mock_datetime.now.return_value = datetime(2020, 4, 27, 20, 16, 1)
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertSetEqual(set(response_json.keys()), {'metrics', 'genesById', 'submissions'})
+        self.assertDictEqual(response_json['metrics'], EXPECTED_MME_DETAILS_METRICS)
+        self.assertEqual(len(response_json['genesById']), 4)
+        self.assertSetEqual(set(response_json['genesById'].keys()), {'ENSG00000233750', 'ENSG00000227232', 'ENSG00000223972', 'ENSG00000186092'})
+        self.assertEqual(len(response_json['submissions']), 3)
+
+    def test_success_story(self):
+        url = reverse(success_story, args=['all'])
+        self.check_staff_login(url)
+
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertListEqual(list(response_json.keys()), ['rows'])
+
+        self.assertEqual(len(response_json['rows']), 2)
+        self.assertDictEqual(response_json['rows'][1], EXPECTED_SUCCESS_STORY)
+
+        url = reverse(success_story, args=['A,T'])
+
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertListEqual(list(response_json.keys()), ['rows'])
+
+        self.assertEqual(len(response_json['rows']), 1)
+        self.assertDictEqual(response_json['rows'][0], EXPECTED_SUCCESS_STORY)
+
+    @mock.patch('seqr.views.apis.summary_data_api.MAX_SAVED_VARIANTS', 1)
+    def test_saved_variants_page(self):
+        url = reverse(saved_variants_page, args=['Tier 1 - Novel gene and phenotype'])
+        self.check_staff_login(url)
+
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()['error'], 'Select a gene to filter variants')
+
+        response = self.client.get('{}?gene=ENSG00000135953'.format(url))
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertSetEqual(set(response_json.keys()), {
+            'projectsByGuid', 'locusListsByGuid', 'savedVariantsByGuid', 'variantFunctionalDataByGuid', 'genesById',
+            'variantNotesByGuid', 'individualsByGuid', 'variantTagsByGuid', 'familiesByGuid'})
+        expected_variant_guids = {
+            'SV0000001_2103343353_r0390_100', 'SV0000007_prefix_19107_DEL_r00', 'SV0000006_1248367227_r0003_tes'}
+        self.assertSetEqual(set(response_json['savedVariantsByGuid'].keys()), expected_variant_guids)
+
+        all_tag_url = reverse(saved_variants_page, args=['ALL'])
+        response = self.client.get('{}?gene=ENSG00000135953'.format(all_tag_url))
+        self.assertEqual(response.status_code, 200)
+        expected_variant_guids.add('SV0000002_1248367227_r0390_100')
+        self.assertSetEqual(set(response.json()['savedVariantsByGuid'].keys()), expected_variant_guids)
